Fixed session cookie on fatal error

+++ sys/php/util.php
@@ -82,7 +82,7 @@
 
         #.ifdef 'HAVE_SESSION_EXTENSION'
     	if(session_id() || vars::$sessionID) {
-    		vars::$Pancake_request->setCookie(session_name(), session_id() ? session_id() : vars::$sessionID, time() + ini_get('session.cookie_lifetime'), ini_get('session.cookie_path'), ini_get('session.cookie_domain'), ini_get('session.cookie_secure'), ini_get('session.cookie_httponly'));
+    		vars::$Pancake_request->setCookie(session_name(), session_id() ? session_id() : vars::$sessionID, time() + ini_get('session.cookie_lifetime'), ini_get('session.cookie_path'), ini_get('session.cookie_domain'), (int) ini_get('session.cookie_secure'), (int) ini_get('session.cookie_httponly'));
     		session_write_close();
 
     		switch(session_cache_limiter()) {
